Make ndims and parent inferrable
Previously, the `@isdefined` conditionals prevented julia from inferring the type of these
methods at compile time. With this change broadcast operations in CUDA.jl like `C .- a'`
can now be fully inferred which was previously impossible because the
BroadcastStyle (which uses ndims internally) could not be decided at compile time.